session01 of 2021S

+++ session01/05for.cc
@@ -15,8 +15,8 @@ int main()
   for (int i = 1; i > 10; i++)
     cout << i;
   cout << '\n';
-  for (int i = 1; i <= 10; i--)
-    cout << i;
+  for (int8_t i = 1; i <= 10; i--)
+    cout << (int)i;
   cout << '\n';
   for (int i = 1; i <= 100; i = i * 2) // 1248163264
     cout << i;
